refactor: fix all C# code quality warnings (53 -> 0)
- SA1210: Sort using directives alphabetically in RecordCampaignCaptureCommandHandler
- SA1201: Move private fields before properties in razor components
- S1905: Remove unnecessary casts in Campaigns and GetCampaignDetailQueryProcessor
- S6608: Use indexing [0] instead of .First() in Campaigns.razor.cs
- S1135: Add pragma directive to suppress intended TODO comment
- Result: Clean build with 0 errors, 0 warnings

+++ src/modules/Spamma.Modules.DomainManagement/Application/QueryProcessors/GetCampaignDetailQueryProcessor.cs
@@ -45,7 +45,7 @@ public class GetCampaignDetailQueryProcessor(IDocumentSession session) : IQueryP
             timeBuckets.Add(new GetCampaignDetailQueryResult.TimeBucket(
                 startTime,
                 endTime,
-                (int)Math.Max(1, campaign.TotalCaptured / 4)));
+                Math.Max(1, campaign.TotalCaptured / 4)));
 
             currentTime = startTime;
         }
